+++ internal/shared/api/routes.go
@@ -37,7 +37,7 @@ func (a *Server) SetupRoutes(r *chi.Mux) {
 	r.Get("/v2/stats", registerHandler(handlers.V2Handler.GetStats))
 	r.Get("/v2/finality-providers", registerHandler(handlers.V2Handler.GetFinalityProviders))
 	r.Get("/v2/global-params", registerHandler(handlers.V2Handler.GetGlobalParams))
-	r.Get("/v2/staker/delegation", registerHandler(handlers.V2Handler.GetDelegationByTxHash))
-	r.Get("/v2/staker/delegations", registerHandler(handlers.V2Handler.GetStakerDelegations))
+	r.Get("/v2/delegation", registerHandler(handlers.V2Handler.GetDelegation))
+	r.Get("/v2/delegations", registerHandler(handlers.V2Handler.GetDelegations))
 	r.Get("/v2/staker/stats", registerHandler(handlers.V2Handler.GetStakerStats))
 }
