+++ internal/v2/service/interface.go
@@ -13,7 +13,7 @@ type V2ServiceProvider interface {
 	SearchFinalityProviders(ctx context.Context, searchQuery string, paginationKey string) ([]*FinalityProviderPublic, string, *types.Error)
 	GetGlobalParams(ctx context.Context) (*GlobalParamsPublic, *types.Error)
 	GetOverallStats(ctx context.Context) (OverallStatsPublic, *types.Error)
-	GetStakerDelegation(ctx context.Context, stakingTxHashHex string) (*StakerDelegationPublic, *types.Error)
-	GetStakerDelegations(ctx context.Context, stakerPKHex string, paginationKey string) ([]*StakerDelegationPublic, string, *types.Error)
+	GetDelegation(ctx context.Context, stakingTxHashHex string) (*StakerDelegationPublic, *types.Error)
+	GetDelegations(ctx context.Context, stakerPKHex string, paginationKey string) ([]*StakerDelegationPublic, string, *types.Error)
 	GetStakerStats(ctx context.Context, stakerPKHex string) (StakerStatsPublic, *types.Error)
 }
